Code rearranged and UI fixes
