--- Geeni-iOS/AddCard.storyboard
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="13196" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="pIY-y1-63W">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13173"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Navigation Controller-->
-        <scene sceneID="17E-qa-fVz">
-            <objects>
-                <navigationController storyboardIdentifier="AddCard" useStoryboardIdentifierAsRestorationIdentifier="YES" id="pIY-y1-63W" sceneMemberID="viewController">
-                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="z5y-rf-BqL">
-                        <rect key="frame" x="0.0" y="20" width="375" height="44"/>
-                        <autoresizingMask key="autoresizingMask"/>
-                    </navigationBar>
-                    <connections>
-                        <segue destination="gwQ-wT-J2u" kind="relationship" relationship="rootViewController" id="d8R-Lt-Wa0"/>
-                    </connections>
-                </navigationController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="1L1-EV-C8N" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="-476" y="95"/>
-        </scene>
-        <!--Add Card View Controller-->
-        <scene sceneID="JOL-AE-jlO">
-            <objects>
-                <viewController id="gwQ-wT-J2u" customClass="AddCardViewController" customModule="Geeni_iOS" customModuleProvider="target" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="9jx-BT-mjd"/>
-                        <viewControllerLayoutGuide type="bottom" id="awr-66-nuk"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="pQj-kp-mIf">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    </view>
-                    <navigationItem key="navigationItem" id="355-sz-g9F">
-                        <barButtonItem key="leftBarButtonItem" image="back" style="plain" id="Meo-n1-e4B">
-                            <button key="customView" opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" id="cM0-le-ZjF">
-                                <rect key="frame" x="16" y="8.5" width="16" height="27"/>
-                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                                <state key="normal" image="back"/>
-                            </button>
-                            <connections>
-                                <action selector="backButtonPressed:" destination="gwQ-wT-J2u" id="gFe-Ml-xHe"/>
-                            </connections>
-                        </barButtonItem>
-                    </navigationItem>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="hQM-Ap-ree" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="390" y="95"/>
-        </scene>
-    </scenes>
-    <resources>
-        <image name="back" width="16" height="27"/>
-    </resources>
-</document>
